cl-merge-pp: Correctly indent printed output.
If the printed values contained strings with newlines, the printed output would
be incorrect, since the indenting was oblivious to the fact that the string
shouldn't be indented.

This means that a structure like this:

(make-person :name "Dave\n")

would be printed as:

(:name "Dave
 ")

Note that there is an extra space at the end of the printed version of the :name
slot which was not originally present.

Signed-off-by: Dan Hackney <dan@haxney.org>